+++ src/wallet.rs
@@ -2,11 +2,11 @@ mod builder;
 mod parser;
 
 pub use builder::signer::{BtcTxSigner, LocalSigner, Wallet};
-#[cfg(feature = "rune")]
-pub use builder::CreateEdictTxArgs;
 pub use builder::{
     CreateCommitTransaction, CreateCommitTransactionArgs, CreateCommitTransactionArgsV2,
     OrdTransactionBuilder, RedeemScriptPubkey, RevealTransactionArgs, ScriptType,
     SignCommitTransactionArgs, TaprootPayload, TxInputInfo, Utxo,
 };
+#[cfg(feature = "rune")]
+pub use builder::{CreateEdictTxArgs, Runestone};
 pub use parser::OrdParser;
